Move Submissions and Feedback log to more general folders
